Use the JavaCompiler execute scripts.
Switched from directly generating JVM byte code to generating
Java source code and compiling it using the javax.tools.JavaCompiler.

This is a little bit of a cheat but allows me to quickly experiment
with the syntax of language without spending a great deal of time
figuring out how to generate the correct byte code.